requirements.txt
feedparser>=6.0.11
openai>=1.109.1
python-dotenv>=1.0.1
yfinance>=0.2.54
